infra/requirements.txt
aws-cdk-lib==2.178.2
aws-cdk.aws-lambda-python-alpha==2.178.2a0